Use the normal Southwest website to log in and fetch reservations
This change was originally done during the migration to the Southwest
web API (in #380). However, I realized that the only change we need to
make to log in and fetch reservations is to use the normal website
instead of the mobile website to do so.

First, the mobile website is loaded to fetch headers to use with the
mobile API in the rest of the script (for now, it seems that the normal
website headers work with the mobile API, but this should hopefully
future-proof it a bit in case that is changed). Then, the normal
website is used to log in and fetch reservations. This allows us to
avoid the redirect from the mobile site to the normal site during login.